+++ src/Phan/Analysis/ParameterTypesAnalyzer.php
@@ -83,8 +83,19 @@ class ParameterTypesAnalyzer
         // are valid
         $is_optional_seen = false;
         foreach ($method->getParameterList() as $i => $parameter) {
-            if ($parameter->getFlags() & Parameter::PARAM_MODIFIER_VISIBILITY_FLAGS) {
+            if ($parameter->getFlags() & Parameter::PARAM_MODIFIER_FLAGS) {
                 if ($method instanceof Method && strcasecmp($method->getName(), '__construct') === 0) {
+                    if ($parameter->getFlags() & ast\flags\MODIFIER_READONLY) {
+                        if (Config::get_closest_minimum_target_php_version_id() < 80100) {
+                            Issue::maybeEmit(
+                                $code_base,
+                                $parameter->createContext($method),
+                                Issue::CompatibleReadonlyProperty,
+                                $parameter->getFileRef()->getLineNumberStart(),
+                                $parameter
+                            );
+                        }
+                    }
                     if (Config::get_closest_minimum_target_php_version_id() < 80000) {
                         Issue::maybeEmit(
                             $code_base,
@@ -102,7 +113,7 @@ class ParameterTypesAnalyzer
                         $parameter->createContext($method),
                         Issue::InvalidNode,
                         $parameter->getFileRef()->getLineNumberStart(),
-                        "Cannot use visibility modifier on parameter $parameter of non-constructor " . $method->getRepresentationForIssue(true)
+                        "Cannot use constructor property promotion modifier on parameter $parameter of non-constructor " . $method->getRepresentationForIssue(true)
                     );
                 }
             }